Refactor commands to separate files and add Sessions command
